fear(react): empty state
