--- packages/react/src/components/blankslate/index.ts
@@ -1,2 +1,0 @@
-export { default as Blankslate } from './components/Blankslate';
-export * from './components/Blankslate';
