+++ packages/react/src/components/empty-state/README.md
@@ -1,0 +1,20 @@
+# Empty State
+
+> Used to provide additional information when no dynamic content exists.
+
+## Usage
+
+To use this component in your app, import as follows:
+
+```jsx
+import { EmptyState } from '@aksara-ui/react';
+
+export default function Component() {
+  return (
+    <EmptyState
+      title="No Project Selected"
+      content="You have not selected a project. Please select one from the dropdown on the left."
+    />
+  );
+}
+```
