+++ packages/react/src/components/empty-state/components/EmptyState.test.tsx
@@ -1,17 +1,17 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
-import Blankslate from './Blankslate';
+import EmptyState from './EmptyState';
 
 describe('components/Blankslate', () => {
   test('renders with title', () => {
-    const { getByText } = render(<Blankslate variant="page" title="Empty Text" content="Blankslate test" />);
+    const { getByText } = render(<EmptyState variant="page" title="Empty Text" content="Blankslate test" />);
 
     const title = getByText('Empty Text');
     expect(title).toBeInTheDocument();
   });
 
   test("doesn't render title in inner variant", () => {
-    const { queryByText } = render(<Blankslate variant="inner" title="Empty Text" content="Blankslate test" />);
+    const { queryByText } = render(<EmptyState variant="inner" title="Empty Text" content="Blankslate test" />);
 
     const title = queryByText(/Empty Text/);
     expect(title).not.toBeInTheDocument();
