+++ packages/react/src/components/empty-state/index.ts
@@ -1,0 +1,2 @@
+export { default as EmptyState } from './components/EmptyState';
+export * from './components/EmptyState';
