+++ packages/react/src/components/index.ts
@@ -1,11 +1,11 @@
 export * from './avatar';
 export * from './badge';
-export * from './blankslate';
 export * from './breadcrumb';
 export * from './button';
 export * from './collapsible';
 export * from './dialog';
 export * from './dropdown';
+export * from './empty-state';
 export * from './form';
 export * from './interactive-card';
 export * from './loading';
